Avoid storing stream type derived from time or range
